Tyr: add a way to import shape in tyr
copy a .poly or .wkt file in the source dir to import it
